fix: 修复ssize_t数据溢出导致打印翻转的问题

+++ interfaces/native/innerkits/src/hitrace_dump.cpp
@@ -667,9 +667,9 @@ void WriteEventFile(std::string &srcPath, int outFd)
         HILOG_ERROR(LOG_CORE, "WriteEventFile: open %{public}s failed.", srcPath.c_str());
         return;
     }
-    ssize_t readLen = 0;
+    int64_t readLen = 0;
     do {
-        ssize_t len = read(srcFd, buffer, PAGE_SIZE);
+        int64_t len = read(srcFd, buffer, PAGE_SIZE);
         if (len <= 0) {
             break;
         }
